LayoutTests/webanimations/animation-creation-basic.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Constructing Animation instances with various arguments.");

debug("The Animation interface is defined.");
shouldBeDefined("Animation");

debug("");
debug("Creating Animation without arguments should have a null timeline.");
const noParameterAnimation = new Animation;
shouldBeNull("noParameterAnimation.timeline");

debug("");
debug("Creating Animation with an undefined argument should have a null timeline.");
const undefinedParameterAnimation = new Animation(undefined);
shouldBeNull("undefinedParameterAnimation.timeline");

debug("");
debug("Creating Animation with a null argument should have a null timeline.");
const nullParameterAnimation = new Animation(null);
shouldBeNull("nullParameterAnimation.timeline");

debug("");
debug("Creating Animation with a timeline has a timeline.");
const documentAnimation = new Animation(document.timeline);
shouldBe("documentAnimation.timeline", "document.timeline");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
