LayoutTests/webanimations/animation-creation-basic-expected.txt

Constructing Animation instances with various arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The Animation interface is defined.
PASS Animation is defined.

Creating Animation without arguments should have a null timeline.
PASS noParameterAnimation.timeline is null

Creating Animation with an undefined argument should have a null timeline.
PASS undefinedParameterAnimation.timeline is null

Creating Animation with a null argument should have a null timeline.
PASS nullParameterAnimation.timeline is null

Creating Animation with a timeline has a timeline.
PASS documentAnimation.timeline is document.timeline

PASS successfullyParsed is true

TEST COMPLETE
